refactor: dev honeycomb events to use lacework api
